requirements.txt
transformers
torch
torchvision
opencv-python
pillow
tqdm
